requirements.txt
attrs
zerorpc
gevent
kids.cache
typing
six
